# specs/003-whoop-dashboard/contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: Whoop Recovery Dashboard — BFF API
  version: "1.0.0"
  description: |
    Internal API exposed by the Express backend-for-frontend (BFF).
    These endpoints are consumed only by the static frontend pages.
    The BFF proxies requests to the Whoop API and manages OAuth tokens.

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /:
    get:
      summary: Landing page
      description: |
        Serves the landing page (`index.html`).
        If valid tokens exist in `.env`, redirects to `/dashboard`.
      responses:
        "200":
          description: Landing page HTML
          content:
            text/html:
              schema:
                type: string
        "302":
          description: Redirect to /dashboard (user already authenticated)
          headers:
            Location:
              schema:
                type: string
                example: /dashboard

  /auth/whoop:
    get:
      summary: Initiate Whoop OAuth flow
      description: |
        Redirects the user to the Whoop authorization screen.
        Generates a random `state` parameter for CSRF protection.
      responses:
        "302":
          description: Redirect to Whoop authorization URL
          headers:
            Location:
              schema:
                type: string
                example: https://api.prod.whoop.com/oauth/oauth2/auth?client_id=...&redirect_uri=...&response_type=code&scope=read:recovery+offline&state=...

  /oauth/callback:
    get:
      summary: OAuth callback handler
      description: |
        Receives the authorization code from Whoop after user authorization.
        Exchanges the code for access + refresh tokens server-side.
        Persists both tokens to `.env`.
        Redirects user to `/dashboard` on success, `/` on failure.
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
          description: Authorization code from Whoop (present on success)
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: CSRF state parameter (must match the value sent in /auth/whoop)
        - name: error
          in: query
          required: false
          schema:
            type: string
          description: Error code if user denied authorization
      responses:
        "302":
          description: Redirect to /dashboard (success) or / (failure)
          headers:
            Location:
              schema:
                type: string
        "400":
          description: Invalid state parameter or missing code
          content:
            text/html:
              schema:
                type: string

  /dashboard:
    get:
      summary: Recovery dashboard page
      description: |
        Serves the dashboard page (`dashboard.html`).
        If no valid tokens exist, redirects to `/`.
      responses:
        "200":
          description: Dashboard page HTML
          content:
            text/html:
              schema:
                type: string
        "302":
          description: Redirect to / (not authenticated)
          headers:
            Location:
              schema:
                type: string
                example: /

  /api/recovery:
    get:
      summary: Get latest recovery data
      description: |
        Proxies the request to Whoop API `GET /developer/v2/recovery?limit=1`.
        Automatically refreshes the access token if expired.
        Returns normalized recovery data or an error object.
      responses:
        "200":
          description: Latest recovery data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryResponse"
        "401":
          description: Not authenticated — tokens missing or refresh failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Whoop API unreachable or returned an error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/status:
    get:
      summary: Authentication status check
      description: |
        Returns whether the user is currently authenticated.
        Does not expose tokens or sensitive data.
      responses:
        "200":
          description: Authentication status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"

components:
  schemas:
    RecoveryResponse:
      type: object
      required:
        - scoreState
      properties:
        scoreState:
          type: string
          enum: [SCORED, PENDING_SCORE, UNSCORABLE]
          description: Whether the recovery has been scored
        recoveryScore:
          type: number
          nullable: true
          description: Recovery percentage (0-100). Null when scoreState is not SCORED.
          example: 44
        hrvRmssdMilli:
          type: number
          nullable: true
          description: Heart rate variability (RMSSD) in milliseconds. Null when not SCORED.
          example: 31.81
        restingHeartRate:
          type: number
          nullable: true
          description: Resting heart rate in beats per minute. Null when not SCORED.
          example: 64
        createdAt:
          type: string
          format: date-time
          nullable: true
          description: When the recovery record was created
          example: "2022-04-24T11:25:44.774Z"

    StatusResponse:
      type: object
      required:
        - authenticated
      properties:
        authenticated:
          type: boolean
          description: Whether valid tokens are present

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: whoop_api_error
        message:
          type: string
          description: Human-readable error message
          example: Could not retrieve recovery data from Whoop. Please try again later.